+++ client/components/Timer.js
@@ -25,9 +25,6 @@ class Timer extends React.Component {
   async componentDidMount() {
     const id = this.props.ticket.id;
     const { data } = await axios.get(`/api/userTickets/${id}/open`);
-    // console.log('CURENTTICKET', currentTicket);
-    // if (data.start) {
-    // }
     if (data.start) {
       const startTime = new Date(data.start);
       const currTime = new Date();
@@ -35,14 +32,12 @@ class Timer extends React.Component {
       // var Seconds_from_T1_to_T2 = diff / 1000;
       // var Seconds_Between_Dates = Math.abs(Seconds_from_T1_to_T2);
       const newDiff = this.millisToMinutesAndSeconds(diff);
-      // console.log('DIFF', diff);
-      // console.log('TIME BETWEEN', newDiff);
       this.setState({
         time: diff
       });
       this.startTimer();
     }
-    setTimeout(console.log(this.state), 1000);
+    // setTimeout(console.log(this.state), 1000);
   }
 
   async startTimer() {
@@ -59,7 +54,6 @@ class Timer extends React.Component {
     );
     this.setState({ status: true });
     const id = this.props.ticket.id;
-    //console.log(this.state.time)
     await axios.post(`/api/userTickets/${id}`);
   }
   async stopTimer() {
@@ -71,7 +65,6 @@ class Timer extends React.Component {
   }
 
   render() {
-    // console.log('ticket', this.props.ticket)
     return (
       <div id="timer">
         <div>{this.millisToMinutesAndSeconds(this.state.time)}</div>
